requirements.txt
jupyter
ipykernel
langchain
langchain-groq
chromadb
langchain_community
pandas
uuid
streamlit
python-dotenv